requirements.txt
flask
flask_socketio
flask_cors
opencv-python
tensorflow
socketio-client
numpy